requirements.txt
# Dependencias Python para Forecasting ML

# Core
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Supabase
supabase>=2.0.0

# Excel
openpyxl>=3.1.0

# Environment
python-dotenv>=1.0.0

# Opcional: Forecasting avanzado (descomentar si usas)
# prophet>=1.1.0
# statsmodels>=0.14.0
# scikit-learn>=1.3.0
